# docs/nextjs-migration/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Coliving Operator API",
    "description": "API for operators to manage property, room, meter, smart door, and contact. All endpoints require **Authorization: Bearer &lt;token&gt;** and **X-API-Username: &lt;username&gt;** headers. You can only access your own client's data.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "https://api.colivingjb.com/api", "description": "Production" }
  ],
  "security": [
    { "BearerAuth": [] },
    { "ApiKeyUsername": [] }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "Your API key (token). Send as: Authorization: Bearer &lt;your-token&gt;"
      },
      "ApiKeyUsername": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Username",
        "description": "Your API username. Get both from portal /docs after your client is approved."
      }
    },
...
  },
  "paths": {
...
    "/smartdoorsetting/list": {
      "post": {
        "summary": "Smart door – List (Read)",
        "tags": ["4) Smart door"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "keyword": { "type": "string" },
                  "propertyId": { "type": "string" },
                  "page": { "type": "integer" },
                  "pageSize": { "type": "integer" }
                }
              },
              "example": { "page": 1, "pageSize": 20 }
            }
          }
        },
        "responses": {
          "200": {
            "description": "items (locks/gateways), total"
          }
        }
      }
    },
    "/smartdoorsetting/get-lock": {
      "post": {
        "summary": "Smart door – Get lock (Read)",
        "tags": ["4) Smart door"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id"],
                "properties": {
                  "id": { "type": "string", "description": "Lock detail ID" }
                }
              },
              "example": { "id": "uuid" }
            }
          }
        },
        "responses": {
          "200": { "description": "Lock object" },
          "404": { "description": "NOT_FOUND" }
        }
      }
    },
    "/smartdoorsetting/get-gateway": {
      "post": {
        "summary": "Smart door – Get gateway (Read)",
        "tags": ["4) Smart door"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id"],
                "properties": {
                  "id": { "type": "string", "description": "Gateway detail ID" }
                }
              },
              "example": { "id": "uuid" }
            }
          }
        },
        "responses": {
          "200": { "description": "Gateway object" },
          "404": { "description": "NOT_FOUND" }
        }
      }
    },
    "/smartdoorsetting/update-lock": {
      "post": {
        "summary": "Smart door – Update lock",
        "tags": ["4) Smart door"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id"],
                "properties": {
                  "id": { "type": "string" },
                  "lockAlias": { "type": "string" },
                  "active": { "type": "boolean" },
                  "childmeter": { "type": "string" }
                }
              },
              "example": { "id": "uuid", "lockAlias": "Room 101" }
            }
          }
        },
        "responses": {
          "200": { "description": "ok: true" }
        }
      }
    },
    "/smartdoorsetting/update-gateway": {
      "post": {
        "summary": "Smart door – Update gateway",
        "tags": ["4) Smart door"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id"],
                "properties": {
                  "id": { "type": "string" },
                  "gatewayName": { "type": "string" }
                }
              },
              "example": { "id": "uuid", "gatewayName": "Gateway 1" }
            }
          }
        },
        "responses": {
          "200": { "description": "ok: true" }
        }
      }
    },
    "/smartdoorsetting/insert-smartdoors": {
      "post": {
        "summary": "Smart door – Add (gateways + locks)",
        "tags": ["4) Smart door"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "gateways": {
                    "type": "array",
                    "items": { "type": "object" }
                  },
                  "locks": {
                    "type": "array",
                    "items": { "type": "object" }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Inserted ids" }
        }
      }
    },
...
  }
}